requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
SQLAlchemy==2.0.23
Werkzeug==2.3.7
WTForms==3.1.1
matplotlib==3.8.0
pandas==2.1.1
python-dotenv==1.0.0
